fix: move all methods of messageSuite to stay together with messageSuite. minor changes/comments
